Refactor out request making methods
